use physically sensible variables
Variables with more physical meaning and more sensible units are
adopted, for example:

- a dimensionless factor for force nonlinearity which gives a good
  measure of the "degree of nonlinearity", instead of a constant of
  dimension 1/[L] which doesn't mean anything...
- using density instead of mass so that fundamental frequency is
  preserved on lattice parameter change, etc.

Variable naming now also follows common physical conventions.

The variables in the updated fput.in are converted from their original
counterparts in a way that the original results are preserved.

Minor refactoring is also done to make code more "object-oriented":
properties one would associate with a string, now actually reside in the
string type! (please, not a string string, but a physical string.)

Function signatures are also changed to better convey their intentions
and reduce code coupling.

With the variable name changes the python part is temporarily broken.

When one starts to write this long a commit message and waffle at the
end(!) it probably means there's too much going on and the programmer
should commit more often. But a lot has happened after the last commit
and I came back to this project to an uncommited mess that I'd left ~2
weeks ago for myself. Really can't get too worked up on it. But hey, I
should do a little write-up at the end.