switch from dt to per_cycle
per_cycle is "number of steps per fundamental period". This ensures the
same "granuarity" of the simulation when one changes the fundamental of
the system.

Of course, if one uses a system with a very long period, the dt will
also become so large that you won't even get one frame per second if you
plot on the "real timescale" of the simulation, but such systems are not
of interest, and there will be hardly any dynamics in that one second
anyhow (so no need for more frames).