add demo video to README
Apparently using <video> tags and linking to raw repo files is no good,
and one has to resort to creating an issue (but not submitting) and
obtaining an assets link by uploading the video there. Although this
works it does mean I will need to keep track of the file and update it
manually if my video changes... sigh.