important integrator bugfix, add 4th order
When I added 4th order integrator coefficients, I was surprised to find
that the chain is not evolved at all. I was even more surprised when,
printing the coefficients, three of the 4th order symp_c and all of the
4th order symp_d coefficients are 0. Guess what? I added an unnecessary
transpose on the coefficient matrices, so instead of coefficients of the
4th order integrator, what was actually used was... well, the 4th
coefficients of the 1st-4th order integrators.

Funnily enough, the third order integrator "worked" before our
discovering this bug. HOW IS THAT POSSIBLE?? Looking at the
coefficients, there happen to be just non-zero element for both the
"transposed" symp_c, and symp_d matrices, so it's probably carrying out
the integration but with the time step distorted in a weird way... or
something. A bit like the first order integrator but with weird
coefficients. If if wasn't for the d_4 = 0 coefficient, I probably
wouldn't even have found out about this. :||||

So much for the transform. Now let's try the third order integrator
again. Yuck... Still wouldn't work. Better read the wiki. Oh... Got two
of the signs wrong.  Haha, easy. Still doesn't work!? Well damn better
read the wiki more carefully. Well didn't see this coming! The
coefficients ought to be iterated through in reverse order, so c_3, d_3
is used first, then c_2, d_2... I remember seeing this very well, but
that's not what's in the code. Maybe I was sleepwalking when I wrote all
this crap...

Anyway that's pretty much it. Afterwards all orders were run and seen to
produce consistent timing results.

In the "evolver" we now use 2x the starting period as patience to
prevent the evolve_while_thresh_met from exiting early, causing the
evolve_outside_thresh to exit in, like a period.